requirements.txt
Flask
Flask-CORS
Pillow
opencv-python
groq
python-dotenv
Werkzeug
